Move tables_manager to utils_db
